refactor: add index files for components and i18n
