+++ apps/client/src/middleware.ts
@@ -1,6 +1,6 @@
 import createMiddleware from 'next-intl/middleware';
 
-import { routing } from './i18n/routing';
+import { routing } from './i18n';
 
 export default createMiddleware(routing);
